Add intro, refs to wikipedia and more on combinators
Also change margins to 1 inch and font size to 11ppt, add references to
the ToC, add basic layout for the rest of the paper and changed foreword
to preface (apparently there's a difference in the two).

I had some trouble getting the references to the wikipedia articles to
work, so it took longer than it should have. I have also been thinking a
lot about the format and structure of the paper and my ultimate goal.

I worked on all this from 15:25-18:20 and 18:30-1930, coming down to
2h55m of work.