Add "A short history" and make minor changes
I worked on this during the following time intervals:

11:50-13:25 -> 1h35m
14:35-17:45 -> 3h10m
18:40-20:20 -> 1h40m

All in all it comes down to 5h25 of work.

I spent a long time on the history, because there was a lot of different
resources and fact checking and I needed to add a lot of books to the
reference. I also read through the rest of the text againt and made some
improvements. The first two blocks of time I spent mostly on writing "A
short history" and the last block of time I spent on getting everything
referenced and making modifications here and there.