Whole lot of work
I have made some progress and thought about structure and ideas for what
I still want to do. I have come with the idea of writing a lisp
interpreter in both C and haskell to compare the paradigms. As I am
writing more an more, it is becoming clearer how to structure things,
because when you haven't written down all the information yet, you are
blindly guessing where you want to take things and how you should
structure it. When you have more info, it is easier to determine all
that. Right now, I'm still racing to get everything down on paper before
the first draft must be handed in. After that, I can think about and
change the structure and language, and finish the interpreters.

oct 31
13:00-18:00 -> start on lisp interpreter in c
19:00-22:00 -> start on parsing and tokenisation and structure of
program

nov 1
10:00-12:00 -> finish tokenisation
13:00-18:00 -> create abstract syntax tree

nov 2
19:00-22:00 -> start on evaluating expressions and representing
functions and other identifiers in an "environment".

nov 6
08:30-09:50 -> iszero and start on subtraction
10:30-10:50 -> pws meeting about idea lisp interpreter and followability
for non-programmers and non-mathematicians
10:50-11:10 -> research subtraction
11:30-12:30 -> continue on subtraction
12:00-13:00 -> restructuring of section 3

nov 18 (today)
15:30-17:30 -> finish up subtraction
18:30-20:10 -> disclaimer on (no) use of symbols, reseach into integers,
rationals, reals, complexes and thinking about the structure and style
of the work and preface.