Nearing closure of using lambda calculus for computation
I have restructured some of the text in this section. I have added stuff
on other kinds of numbers, but I still need to complete it. I also still
need to add to the boolean comparison subsection, because it is (very)
lacking. Furthermore, the combinatory logic section is still lacking a
bit, because there are some claims made or things told that aren't
elaborated on any further.

As for my plans from now on: I am trying to get all the information I
need down on paper, I am sort of rushing it, it isn't pretty. Later on I
will go back and rewrite a lot, if not most, of the text. Most of the
explainations are currently very lacking or vague. They merely get some
idea across, but aren't very clear or structured at all. I will need to
restructure it and maybe add some graphics that make it easier to
visualise what is being talked about. Most of the purely theoretical
stuff is getting to a close, so I am hoping to be able to get to it
quickly.

I have worked from 18:45 to 21:00, coming down to 2h15m.